AImodel/pyproject.toml
[project]
name = "aimodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = ["aiohappyeyeballs (==2.4.4)", "aiohttp (==3.11.11)", "aiosignal (==1.3.2)", "annotated-types (==0.7.0)", "anyio (==4.8.0)", "backcall (==0.2.0)", "charset-normalizer (==3.4.1)", "dataclasses-json (==0.6.7)", "distro (==1.9.0)", "docopt (==0.6.2)", "frozenlist (==1.5.0)", "greenlet (==3.1.1)", "httpcore (==1.0.7)", "httpx (==0.28.1)", "httpx-sse (==0.4.0)", "idna (==3.10)", "ipython (==8.31.0)", "jiter (==0.8.2)", "jsonpatch (==1.33)", "jsonpointer (==3.0.0)", "langchain (==0.3.14)", "langchain-core (==0.3.30)", "langchain-openai (==0.3.0)", "langchain-text-splitters (==0.3.5)", "langsmith (==0.2.11)", "marshmallow (==3.25.1)", "multidict (==6.1.0)", "mypy-extensions (==1.0.0)", "numpy (==2.0.0)", "openai (==1.59.8)", "orjson (==3.10.14)", "pickleshare (==0.7.5)", "propcache (==0.2.1)", "pydantic (==2.10.5)", "pydantic-core (==2.27.2)", "pydantic-settings (==2.7.1)", "python-dotenv (==1.0.1)", "regex (==2024.11.6)", "requests-toolbelt (==1.0.0)", "sniffio (==1.3.1)", "sqlalchemy (==2.0.37)", "tenacity (==8.5.0)", "tiktoken (==0.8.0)", "tqdm (==4.67.1)", "typing-inspect (==0.9.0)", "uv (==0.5.21)", "yarg (==0.1.9)", "yarl (==1.18.3)"]
